sidbi specific changes

+++ service-loans/src/main/java/com/capitaworld/service/loans/service/fundseeker/corporate/impl/FsPastPerformanceDetailServiceImpl.java
@@ -90,8 +90,8 @@ public class FsPastPerformanceDetailServiceImpl implements FsPastPerformanceDeta
 				FsPastPerformanceDetails fsPastPerformanceDetails1 = new FsPastPerformanceDetails();
 				OperatingStatementDetails operatingStatementPast1YearDetails = operatingStatementDetailsRepository.getOperatingStatementDetailsByAppIdAndFinYear(applicationId, pastYear1+"");
 				if (!CommonUtils.isObjectNullOrEmpty(operatingStatementPast1YearDetails)) {
-					fsPastPerformanceDetails1.setNetSalesPastYear1(operatingStatementPast1YearDetails.getNetSales().longValue());
-					fsPastPerformanceDetails1.setNetProfitPastYear1(operatingStatementPast1YearDetails.getNetProfitOrLoss().longValue());
+					fsPastPerformanceDetails1.setNetSalesPastYear1(operatingStatementPast1YearDetails.getNetSales().doubleValue());
+					fsPastPerformanceDetails1.setNetProfitPastYear1(operatingStatementPast1YearDetails.getNetProfitOrLoss().doubleValue());
 				}else{
 					logger.error("Error while getting net sales/net profit loss past 1 year details from operating statement");
 				}
@@ -99,22 +99,22 @@ public class FsPastPerformanceDetailServiceImpl implements FsPastPerformanceDeta
 
 				OperatingStatementDetails operatingStatementPast2YearDetails = operatingStatementDetailsRepository.getOperatingStatementDetailsByAppIdAndFinYear(applicationId, pastYear2+"");
 				if (!CommonUtils.isObjectNullOrEmpty(operatingStatementPast2YearDetails)) {
-					fsPastPerformanceDetails1.setNetSalesPastYear2(operatingStatementPast2YearDetails.getNetSales().longValue());
-					fsPastPerformanceDetails1.setNetProfitPastYear2(operatingStatementPast2YearDetails.getNetProfitOrLoss().longValue());
+					fsPastPerformanceDetails1.setNetSalesPastYear2(operatingStatementPast2YearDetails.getNetSales().doubleValue());
+					fsPastPerformanceDetails1.setNetProfitPastYear2(operatingStatementPast2YearDetails.getNetProfitOrLoss().doubleValue());
 				}else{
 					logger.error("Error while getting net sales/net profit loss past 2 year details from operating statement");
 				}
 
 				LiabilitiesDetails liabilitiesPast1YearDetails =  liabilitiesDetailsRepository.getLiabilitiesDetailsByAppIdAndFinYear(applicationId,pastYear1+"");
 				if (!CommonUtils.isObjectNullOrEmpty(liabilitiesPast1YearDetails)) {
-					fsPastPerformanceDetails1.setCompNetWorthPastYear1(liabilitiesPast1YearDetails.getNetWorth().longValue());
+					fsPastPerformanceDetails1.setCompNetWorthPastYear1(liabilitiesPast1YearDetails.getNetWorth().doubleValue());
 				}else{
 					logger.error("Error while getting company net worth past 1 year details from liabilities statement");
 				}
 
 				LiabilitiesDetails liabilitiesPast2YearDetails =  liabilitiesDetailsRepository.getLiabilitiesDetailsByAppIdAndFinYear(applicationId,pastYear2+"");
 				if (!CommonUtils.isObjectNullOrEmpty(liabilitiesPast2YearDetails)) {
-					fsPastPerformanceDetails1.setCompNetWorthPastYear2(liabilitiesPast2YearDetails.getNetWorth().longValue());
+					fsPastPerformanceDetails1.setCompNetWorthPastYear2(liabilitiesPast2YearDetails.getNetWorth().doubleValue());
 				}else{
 					logger.error("Error while getting company net worth past 2 year details from liabilities statement");
 				}
